update to internal InProcessTaskEngine
