+++ strands-ts/src/background-tasks/in-process/__tests__/engine-transitions.test.ts
@@ -7,11 +7,11 @@ import {
   shutdownEngines,
   waitForStatus,
 } from './engine-test-helpers.js'
-import type { TestContext, TestOutcome } from './engine-test-helpers.js'
+import type { TestOutcome } from './engine-test-helpers.js'
 
 afterEach(shutdownEngines)
 
-describe('BackgroundTaskEngine', () => {
+describe('InProcessTaskEngine', () => {
   describe('waiting and transitions', () => {
     it('cancels observations without cancelling work', async () => {
       const finish = deferred<TestOutcome>()
@@ -34,15 +34,13 @@ describe('BackgroundTaskEngine', () => {
       await engine.wait(task.taskId)
     })
 
-    it('pauses, persists state, and resumes the same attempt', async () => {
-      const contexts: TestContext[] = []
+    it('pauses, persists state, and resumes execution', async () => {
       const engine = initialize(
-        createEngine(async (context) => {
-          contexts.push(context)
-          return context.state
-            ? { status: 'completed', result: { value: 'resumed' } }
+        createEngine(async ({ state }) =>
+          state
+            ? { status: 'completed', result: { value: state.phase } }
             : { status: 'paused', state: { phase: 'waiting' } }
-        })
+        )
       )
       const task = engine.submit({ descriptor: { value: 'work' } })
       const paused = await engine.wait(task.taskId)
@@ -50,36 +48,35 @@ describe('BackgroundTaskEngine', () => {
         status: 'paused',
         state: { phase: 'waiting' },
       })
-      expect(() =>
-        engine.resume(task.taskId, () => ({ state: undefined as unknown as { phase: string }, ready: false }))
-      ).toThrow('task.state is required while paused')
       expectTask(
         engine.resume(task.taskId, () => ({ state: { phase: 'still waiting' }, ready: false })),
         task,
         { status: 'paused', state: { phase: 'still waiting' } }
       )
-      expect(contexts).toHaveLength(1)
       engine.resume(task.taskId, (state) => ({ state, ready: true }))
       expectTask(await engine.wait(task.taskId), task, {
         status: 'completed',
-        result: { value: 'resumed' },
+        result: { value: 'still waiting' },
       })
-      expect(contexts[0]!.attemptId).toBe(contexts[1]!.attemptId)
-      expect(contexts[0]!.executionId).not.toBe(contexts[1]!.executionId)
     })
 
     it('cancels running work, wakes waiters, and removes after execution settles', async () => {
       const finish = deferred<TestOutcome>()
-      const events: string[] = []
-      const engine = initialize(createEngine(async () => finish.promise, { onEvent: ({ type }) => events.push(type) }))
+      let executionSignal: AbortSignal | undefined
+      const engine = initialize(
+        createEngine(async ({ cancelSignal }) => {
+          executionSignal = cancelSignal
+          return finish.promise
+        })
+      )
       const task = engine.submit({ descriptor: { value: 'work' } })
       await waitForStatus(engine, task.taskId, 'working')
       const waiting = engine.wait(task.taskId)
 
       engine.cancel(task.taskId, { reason: 'Stop work' })
+      expect(executionSignal?.reason).toBe('Stop work')
       expectTask(await waiting, task, {
         status: 'cancelled',
-        cancellationReason: 'Stop work',
       })
       engine.remove(task.taskId)
       expect(engine.get(task.taskId)).toBeUndefined()
@@ -87,12 +84,34 @@ describe('BackgroundTaskEngine', () => {
         `Background task '${task.taskId}' was not found`
       )
       finish.resolve({ status: 'completed', result: { value: 'late' } })
-      await engine.shutdown({ mode: 'drain', timeout: 1_000 })
-      expect(events).toEqual(['admitted', 'executionStarted', 'cancelled', 'executionFinished'])
-      expect(engine.get(task.taskId)).toBeUndefined()
+      await engine.shutdown({ timeout: 1_000 })
     })
 
-    it('rejects invalid configuration and record values without stopping work', async () => {
+    it('cancels queued work without executing it', async () => {
+      const finish = deferred<TestOutcome>()
+      const executions: string[] = []
+      const engine = initialize(
+        createEngine(
+          async ({ descriptor }) => {
+            executions.push(descriptor.value)
+            return finish.promise
+          },
+          { maxConcurrency: 1 }
+        )
+      )
+      const running = engine.submit({ descriptor: { value: 'running' } })
+      await waitForStatus(engine, running.taskId, 'working')
+      const queued = engine.submit({ descriptor: { value: 'queued' } })
+
+      expectTask(engine.cancel(queued.taskId, { reason: 'No longer needed' }), queued, {
+        status: 'cancelled',
+      })
+      finish.resolve({ status: 'completed', result: { value: 'done' } })
+      await engine.waitForIdle()
+      expect(executions).toEqual(['running'])
+    })
+
+    it('rejects invalid timer configuration without stopping work', async () => {
       const complete = async (): Promise<TestOutcome> => ({ status: 'completed', result: { value: 'done' } })
       expect(() => createEngine(complete, { maxConcurrency: 0 })).toThrow('maxConcurrency must be a positive')
       expect(() => createEngine(complete, { timeout: 0 })).toThrow('timeout must be a positive')
@@ -101,22 +120,12 @@ describe('BackgroundTaskEngine', () => {
 
       const finish = deferred<TestOutcome>()
       const engine = initialize(createEngine(async () => finish.promise))
-      expect(() => engine.submit({ descriptor: { value: 'invalid' }, idempotencyKey: '' })).toThrow(
-        'task.idempotencyKey must be a non-empty string'
-      )
       const task = engine.submit({ descriptor: { value: 'work' } })
       await waitForStatus(engine, task.taskId, 'working')
-      expect(() => engine.cancel(task.taskId, { reason: '' })).toThrow(
-        'task.cancellationReason must be a non-empty string'
-      )
       finish.resolve({ status: 'completed', result: { value: 'done' } })
       await engine.wait(task.taskId)
-      await expect(engine.shutdown({ mode: 'drain', timeout: 0 })).rejects.toThrow(
-        'shutdown timeout must be a positive'
-      )
-      await expect(engine.shutdown({ mode: 'drain', timeout: 2 ** 31 })).rejects.toThrow(
-        'shutdown timeout must be at most'
-      )
+      await expect(engine.shutdown({ timeout: 0 })).rejects.toThrow('shutdown timeout must be a positive')
+      await expect(engine.shutdown({ timeout: 2 ** 31 })).rejects.toThrow('shutdown timeout must be at most')
     })
   })
 })
